pyproject.toml
[project]
name = "localtalk"
version = "0.1.0-alpha.2"
description = "A local/offline-capable voice assistant with speech recognition, LLM processing, and text-to-speech"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Anthony Wu", email = "pls-file-gh-issue@users.noreply.github.com"}
]
license = {text = "MIT"}
keywords = ["voice-assistant", "tts", "speech-recognition", "llm", "offline", "apple-silicon", "privacy", "local"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Environment :: MacOS X",
]

dependencies = [
    "mlx>=0.27.1",
    "mlx-audio>=0.2.3",
    "mlx-lm>=0.26.1",
    "mlx-vlm>=0.3.2",
    "misaki[en,ja,ko,zh]>=0.9.4",
    "numpy>=2.0,<2.3.0",  # restricted by opencv-python
    "openai-whisper>=20250625",
    "pydantic>=2.0.0",
    "pip>25",  # kokoro pipeline downloads uses this, anti-pattern but we'll deal with it for now
    "rich>=14.0.0",
    "sounddevice>=0.5",
    "soundfile>=0.13",
    "transformers>=4.54.0",
]

[project.optional-dependencies]
chatterbox = [
    "chatterbox-tts>=0.1.0",
    "nltk>=3.9.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
localtalk = "localtalk.cli:main"

[project.urls]
Homepage = "https://github.com/anthonywu/localtalk"
Documentation = "https://github.com/anthonywu/localtalk#readme"
Repository = "https://github.com/anthonywu/localtalk.git"
Issues = "https://github.com/anthonywu/localtalk/issues"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
fix = true
indent-width = 4
line-length = 120
respect-gitignore = true
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --strict-markers --cov=localtalk"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]
